+++ .gitignore
@@ -5,4 +5,4 @@ build/
 node_modules/
 .idea/
 demos/local/
-demos/light-signup-demo.html
+demos/email-only-signup-demo.html
