+++ demos/src/config.json
@@ -4,7 +4,7 @@
   },
   "demos": [
     {
-      "name": "light-signup-demo",
+      "name": "email-only-signup-demo",
       "sass": "demos/src/demo.scss",
       "template": "demos/src/demo.html",
       "js": "demos/src/demo.js"
