+++ main.scss
@@ -1,1 +1,1 @@
-@import "./src/light-signup";
+@import "./src/email-only-signup";
